Reorganiza os models.
